pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "Slurm-web"
version = "6.0.0"
description = "Web interface to Slurm HPC clusters"
license = {text = "MIT"}
requires-python = ">=3.6"
keywords = ["slurm", "hpc", "cluster"]
authors = [
    {name = "Rémi Palancher", email = "remi@rackslab.io"},
]
dependencies = [
    "alembic",
    "aiohttp",
    "Flask",
    "psycopg2-binary",
    "RFL.authentication >= 1.5.0",
    "RFL.core >= 1.1.0",
    "RFL.log >= 1.4.0",
    "RFL.settings >= 1.4.0",
    "RFL.web >= 1.3.0",
    "SQLAlchemy",
    "setuptools"
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "Flask-Cors",
    "sshtunnel",
]
agent = [
    "ClusterShell",
    "prometheus-client",
    "RacksDB[web]",
    "redis",
    "requests",
]
gateway = [
    "markdown",
]
tests = [
    "coverage",
    "Jinja2",
    "parameterized",
    "pytest",
    "pytest-cov",
]

[project.scripts]
slurm-web = 'slurmweb.exec.main:SlurmwebExecMain.run'

[tool.setuptools.packages.find]
include = ['slurmweb*']

[tool.setuptools.package-data]
slurmweb = [
    "alembic.ini",
    "alembic/*.py",
    "alembic/*.mako",
    "alembic/README",
    "alembic/versions/*.py",
]

[project.urls]
"Homepage" = "https://github.com/rackslab/slurm-web"
"Bug Tracker" = "https://github.com/rackslab/slurm-web/issues"

[tool.ruff.lint]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
